+++ tog_headlines_template.rb
@@ -1,5 +1,4 @@
 plugin 'fckeditor', :git => "git://github.com/molpe/fckeditor.git"
-plugin 'seo_urls', :svn => "http://svn.redshiftmedia.com/svn/plugins/seo_urls"
 
 plugin 'tog_headlines', :git => "git://github.com/tog/tog_headlines.git"